fix(search): extract getImageUrl to lib/api/tmdb-images.ts so client bundle skips pino/fs (#101)
